Update tgsjo.ijs
(1) When reporting an opengl error, shut down all running timers and throw an error in J. This should allow faster isolation of problems (though it is a bit ugly).

(2) In several cases 10 was used to refer to the compiled opengl program -- this is unreliable. Instead, the value of sprog should be used. (It's 10, when it needs to be 10, but may take on other values.)